requirements.txt
Django==1.10.5
django-allauth==0.30.0
django-cors-headers==2.0.0
django-rest-auth==0.9.0
djangorestframework==3.5.3
oauthlib==2.0.1
psycopg2==2.6.2
python-openid==2.2.5
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0